+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## $v
+## v160721.58752
 
 - Updating to the latest websharks/core.
 - Adding Jetpack utils for Markdown parsing.